+++ tests/CodeListControllerTest.cs
@@ -70,13 +70,13 @@ public class CodeListControllerTest
     [TestMethod]
     public async Task GetHydrotestCodesByTestKind()
     {
-        var codeListsInfiltrationEssay = await controller.GetAsync("", 15203179);
+        var codeListsInfiltrationEssay = await controller.GetAsync("", new int[] { 15203179 });
         Assert.AreEqual(6, codeListsInfiltrationEssay.Count());
         Assert.IsTrue(codeListsInfiltrationEssay.All(c => c.Schema == HydrogeologySchemas.FlowdirectionSchema
                                                        || c.Schema == HydrogeologySchemas.EvaluationMethodSchema
                                                        || c.Schema == HydrogeologySchemas.HydrotestResultParameterSchema));
 
-        var codeListsPumpEssay = await controller.GetAsync("", 15203170);
+        var codeListsPumpEssay = await controller.GetAsync("", new int[] { 15203170 });
         Assert.AreEqual(13, codeListsPumpEssay.Count());
         Assert.IsTrue(codeListsPumpEssay.All(c => c.Schema == HydrogeologySchemas.FlowdirectionSchema
                                                 || c.Schema == HydrogeologySchemas.EvaluationMethodSchema
@@ -86,7 +86,7 @@ public class CodeListControllerTest
     [TestMethod]
     public async Task GetHydrotestCodesByInexistantTestKind()
     {
-        var codeListsInfiltrationEssay = await controller.GetAsync("", 6715779);
+        var codeListsInfiltrationEssay = await controller.GetAsync("", new int[] { 6715779 });
         Assert.AreEqual(0, codeListsInfiltrationEssay.Count());
     }