+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 ### Added
 
 - Lithology csv import.
+- Table and controller for lithostratigraphy.
 
 ### Changed
 
+- The hydrotest input mask now supports selecting multiple hydrotest kinds.
 - Enable all file types as borehole attachments.
 
 ### Fixed